requirements.txt
# Brain MRI Diagnostic System — Dependencies
torch>=2.0.0
torchvision>=0.15.0
gradio>=4.0.0
Pillow>=9.0.0
numpy>=1.24.0
reportlab>=4.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
